+++ .dialyzer_ignore.exs
@@ -1,6 +1,1 @@
-[
-  {":0:unknown_function Function :persistent_term.get/0 does not exist."},
-  {":0:unknown_function Function :persistent_term.get/2 does not exist."},
-  {":0:unknown_function Function :persistent_term.put/2 does not exist."},
-  {":0:unknown_function Function :persistent_term.erase/1 does not exist."}
-]
+[]