+++ bench/benchee_helper.exs
@@ -2,13 +2,9 @@ alias TimeZoneInfo.{DataStore, ExternalTermFormat}
 
 Application.ensure_all_started(:tzdata)
 
-Code.require_file("test/support/time_zone_info/data_store/server.exs")
-
 {:ok, data} = "priv/data.etf" |> File.read!() |> ExternalTermFormat.decode()
 
-DataStore.PersistentTerm.put(data)
-DataStore.ErlangTermStorage.put(data)
-DataStore.Server.put(data)
+DataStore.put(data)
 
 BencheeDsl.run(
   time: 10,