+++ lib/time_zone_info/data_store.ex
@@ -1,132 +1,118 @@
 defmodule TimeZoneInfo.DataStore do
   @moduledoc """
-  A behaviour to store data and serve them later on.
+  A module to store the time zone data in the persistent term.
   """
 
-  @default_time_zone "Etc/UTC"
+  @key :time_zone_info
 
-  @doc "Puts the given `data` into the store."
-  @callback put(data :: TimeZoneInfo.data()) :: :ok | :error
+  @spec put(data :: TimeZoneInfo.data()) :: :ok | :error
+  def put(data) do
+    :persistent_term.put({@key, :time_zones}, Map.get(data, :time_zones, %{}))
+    :persistent_term.put({@key, :rules}, Map.get(data, :rules, %{}))
+    :persistent_term.put({@key, :links}, Map.get(data, :links, %{}))
+    :persistent_term.put({@key, :config}, Map.get(data, :config))
+    :persistent_term.put({@key, :version}, Map.get(data, :version))
+  end
 
-  @doc """
-  Returns the `transitions` for a given `time_zone`.
-  """
-  @callback fetch_transitions(time_zone :: Calendar.time_zone()) ::
-              {:ok, transitions :: [TimeZoneInfo.transition()]} | {:error, :transitions_not_found}
-
-  @doc """
-  Returns `rules` for a given `rule_name`.
-  """
-  @callback fetch_rules(rule_name :: TimeZoneInfo.rule_name()) ::
-              {:ok, rules :: [TimeZoneInfo.rule()]} | {:error, :rule_not_found}
-
-  @doc """
-  Returns the list of all available time zones with or without links. The option
-  `:links` can be used to customize the list.
-
-  Values for `:links`:
-  - `:ignore` just the time zone names will be returned
-  - `:only` just the link name will be returned
-  - `:include` the time zone and link names will be returned
-
-  The list will be sorted.
-  """
-  @callback get_time_zones(links: :ignore | :only | :include) :: [Calendar.time_zone()]
-
-  @doc """
-  Returns true if the `DataSore` is empty.
-  """
-  @callback empty? :: boolean
-
-  @doc """
-  Returns the version of the IANA database from which the data was generated.
-  """
-  @callback version :: String.t() | nil
-
-  @doc """
-  Deletes all data in the data store.
-  """
-  @callback delete! :: :ok
-
-  @doc """
-  Returns infos about the data store.
-  """
-  @callback info :: term()
-
-  @optional_callbacks info: 0
-
-  # Implementation
-
-  defp impl do
-    case Application.fetch_env!(:time_zone_info, :data_store) do
-      :detect -> detect()
-      module -> module
+  @spec fetch_transitions(time_zone :: Calendar.time_zone()) ::
+          {:ok, transitions :: [TimeZoneInfo.transition()]} | {:error, :transitions_not_found}
+  def fetch_transitions(time_zone, is_link \\ false) do
+    with :error <- fetch(:time_zones, time_zone) do
+      fetch_transitions_by_link(time_zone, is_link)
     end
   end
 
-  defp detect do
-    module =
-      case function_exported?(:persistent_term, :get, 0) do
-        true -> __MODULE__.PersistentTerm
-        false -> __MODULE__.ErlangTermStorage
-      end
-
-    Application.put_env(:time_zone_info, :data_store, module)
-
-    module
-  end
-
-  @doc false
-  @spec fetch_transitions(Calendar.time_zone()) ::
-          {:ok, [TimeZoneInfo.transition()]} | {:error, :transitions_not_found}
-  def fetch_transitions(time_zone), do: impl().fetch_transitions(time_zone)
-
-  @doc false
-  @spec get_time_zones(links: :ignore | :only | :include) :: [Calendar.time_zone()]
-  def get_time_zones(opts) do
-    time_zones = impl().get_time_zones(opts)
-
-    case opts[:links] do
-      :only ->
-        time_zones
-
-      _else ->
-        case Enum.member?(time_zones, @default_time_zone) do
-          true -> time_zones
-          false -> Enum.sort([@default_time_zone | time_zones])
-        end
+  @spec fetch_rules(rule_name :: TimeZoneInfo.rule_name()) ::
+          {:ok, rules :: [TimeZoneInfo.rule()]} | {:error, :rules_not_found}
+  def fetch_rules(rules) do
+    with :error <- fetch(:rules, rules) do
+      {:error, :rules_not_found}
     end
   end
 
-  @doc false
-  @spec fetch_rules(TimeZoneInfo.rule_name()) ::
-          {:ok, [TimeZoneInfo.rule()]} | {:error, :rules_not_found}
-  def fetch_rules(rule_name), do: impl().fetch_rules(rule_name)
+  @spec get_time_zones(links: :include | :ignore | :only) :: [Calendar.time_zone()]
+  def get_time_zones(links: :ignore) do
+    case fetch(:time_zones) do
+      {:ok, time_zones} -> time_zones |> Map.keys() |> Enum.sort()
+      :error -> []
+    end
+  end
 
-  @doc false
-  @spec put(TimeZoneInfo.data()) :: :ok | :error
-  def put(data), do: impl().put(data)
+  def get_time_zones(links: :only) do
+    case fetch(:links) do
+      {:ok, links} -> links |> Map.keys() |> Enum.sort()
+      :error -> []
+    end
+  end
 
-  @doc false
-  @spec empty? :: boolean()
-  def empty?, do: impl().empty?()
+  def get_time_zones(links: :include) do
+    with {:ok, time_zones} <- fetch(:time_zones),
+         {:ok, links} <- fetch(:links) do
+      Enum.sort(Map.keys(time_zones) ++ Map.keys(links))
+    else
+      _error -> []
+    end
+  end
 
-  @doc false
+  @spec empty? :: boolean
+  def empty?, do: :persistent_term.get({@key, :time_zones}, nil) == nil
+
   @spec version :: String.t() | nil
-  def version, do: impl().version()
+  def version, do: get(:version)
 
-  @doc false
   @spec delete! :: :ok
-  def delete!, do: impl().delete!()
+  def delete! do
+    :persistent_term.erase({@key, :time_zones})
+    :persistent_term.erase({@key, :rules})
+    :persistent_term.erase({@key, :links})
+    :persistent_term.erase({@key, :config})
+    :persistent_term.erase({@key, :version})
+    :ok
+  end
 
-  @doc false
   @spec info :: term()
   def info do
-    impl = impl()
+    memory =
+      [:time_zones, :rules, :links, :config, :version]
+      |> Enum.map(fn sub_key ->
+        case :persistent_term.get({@key, sub_key}, nil) do
+          nil -> 0
+          data -> data |> :erlang.term_to_binary() |> byte_size()
+        end
+      end)
+      |> Enum.sum()
 
-    case function_exported?(impl, :info, 0) do
-      false -> :no_implementation_found
-      true -> impl.info()
+    %{
+      version: version(),
+      memory: memory,
+      time_zones: length(get_time_zones(links: :ignore)),
+      links: length(get_time_zones(links: :only))
+    }
+  end
+
+  defp get(key) do
+    :persistent_term.get({@key, key})
+  end
+
+  defp fetch(key) do
+    case :persistent_term.get({@key, key}, :error) do
+      :error -> :error
+      value -> {:ok, value}
+    end
+  end
+
+  defp fetch(key, sub_key) do
+    with {:ok, map} <- fetch(key) do
+      Map.fetch(map, sub_key)
+    end
+  end
+
+  defp fetch_transitions_by_link(_link, true), do: {:error, :transitions_not_found}
+
+  defp fetch_transitions_by_link(link, false) do
+    case fetch(:links, link) do
+      {:ok, time_zone} -> fetch_transitions(time_zone, true)
+      :error -> {:error, :transitions_not_found}
     end
   end
 end